test: consolidate file language defaults coverage
